+++ src/tooltip/Tooltip.tsx
@@ -1,21 +1,13 @@
-import React, {
-  forwardRef,
-  useState,
-  useEffect,
-  useRef,
-  useImperativeHandle,
-  cloneElement,
-  isValidElement,
-} from 'react';
+import React, { forwardRef, useState, useEffect, useRef, useImperativeHandle } from 'react';
 import classNames from 'classnames';
-import Popup, { PopupVisibleChangeContext } from '../popup';
+import Popup, { PopupVisibleChangeContext, PopupRef } from '../popup';
 import useConfig from '../hooks/useConfig';
 import { TdTooltipProps } from './type';
 import { tooltipDefaultProps } from './defaultProps';
 
 export type TooltipProps = TdTooltipProps;
 
-interface RefProps {
+export interface TooltipRef extends PopupRef {
   setVisible?: (v: boolean) => void;
 }
 
@@ -33,7 +25,7 @@ const Tooltip = forwardRef((props: TdTooltipProps, ref) => {
   const { classPrefix } = useConfig();
   const [isTipShowed, setTipshow] = useState(duration !== 0);
   const [timeup, setTimeup] = useState(false);
-  const popupRef = useRef<HTMLDivElement>();
+  const popupRef = useRef<PopupRef>();
   const timerRef = useRef<number | null>(null);
   const [offset, setOffset] = useState([0, 0]);
   const toolTipClass = classNames(
@@ -58,22 +50,6 @@ const Tooltip = forwardRef((props: TdTooltipProps, ref) => {
       x,
       y,
     };
-  };
-
-  const getTriggerChildren = (children) => {
-    const displayName = children.type?.displayName;
-    // disable情况下button不响应mouse事件，但需要展示tooltip，所以要包裹一层
-    if ((children.type === 'button' || displayName === 'Button') && children?.props?.disabled) {
-      const displayStyle = children.props?.style?.display ? children.props.style.display : 'inline-block';
-      const child = cloneElement(children, {
-        style: {
-          ...children.props.style,
-          pointerEvents: 'none',
-        },
-      });
-      return <span style={{ display: displayStyle, cursor: 'not-allowed' }}>{child}</span>;
-    }
-    return children;
   };
 
   const handleShowTip = (visible: boolean, { e, trigger }: PopupVisibleChangeContext) => {
@@ -102,12 +78,10 @@ const Tooltip = forwardRef((props: TdTooltipProps, ref) => {
     };
   }, [duration, timeup]);
 
-  useImperativeHandle(
-    ref,
-    (): RefProps => ({
-      setVisible,
-    }),
-  );
+  useImperativeHandle(ref, () => ({
+    setVisible,
+    ...((popupRef.current || {}) as any),
+  }));
 
   return (
     <Popup
@@ -117,11 +91,13 @@ const Tooltip = forwardRef((props: TdTooltipProps, ref) => {
       overlayClassName={toolTipClass}
       visible={isTipShowed}
       onVisibleChange={handleShowTip}
-      popperModifiers={isPlacedByMouse ? [{ name: 'offset', options: { offset } }] : []}
+      popperOptions={{
+        modifiers: isPlacedByMouse ? [{ name: 'offset', options: { offset } }] : [],
+      }}
       placement={isPlacedByMouse ? 'bottom-left' : placement}
       {...restProps}
     >
-      {isValidElement(children) ? getTriggerChildren(children) : children}
+      {children}
     </Popup>
   );
 });