+++ src/upload/__tests__/upload.test.tsx
@@ -1,4 +1,6 @@
 import { testExamples } from '@test/utils';
+import MockDate from 'mockdate';
 
+MockDate.set('2022-08-27');
 // 测试组件代码 Example 快照
 testExamples(__dirname);