+++ src/upload/_example/custom-drag.jsx
@@ -33,9 +33,9 @@ export default function CustomDrag() {
 
       function renderFiles(files) {
         return (
-          <ul>
+          <ul style={{ padding: '0' }}>
             {files.map((file) => (
-              <li key={file.name}>{file.name}</li>
+              <li key={file.name} style={{ listStyleType: 'none' }}>{file.name}</li>
             ))}
           </ul>
         );
@@ -45,7 +45,7 @@ export default function CustomDrag() {
         <>
           {files?.length ? renderFiles(files) : renderCustomDrag()}
           {files?.length > 0 && (
-            <Button variant="base" style={{ marginTop: '36px' }}>
+            <Button variant="base" style={{ marginTop: '16px' }}>
               更换文件
             </Button>
           )}
@@ -58,8 +58,7 @@ export default function CustomDrag() {
   );
   return (
     <Space direction="vertical">
-      <Button variant="outline" onClick={upload}>
-        <CloudUploadIcon />
+      <Button variant="outline" icon={<CloudUploadIcon />} onClick={upload}>
         点击上传
       </Button>
       <Upload
@@ -68,7 +67,7 @@ export default function CustomDrag() {
         action="https://service-bv448zsw-1257786608.gz.apigw.tencentcs.com/api/upload-demo"
         draggable
         theme="custom"
-        customDraggerRender={customDraggerRender}
+        dragContent={customDraggerRender}
         onChange={handleChange}
         onTrigger={handleChange}
         onFail={handleFail}